Added functions RequestActor and CheckInactiveActors. RequestActor revieves a vector of std::unique_ptr<IComponent> and activates a corresponding set of components if there are any available actors. CheckInactiveActors will identify any newly inactivated Actors and place them at the back of the set of active Actors. This way the components are alwayes sorted by 'activity', keeping the active Actors at the first half of the component collections.

+++ TwinStick/ActorCollection.h
@@ -10,8 +10,8 @@
 
 struct ActorCollection
 {
-	std::vector<bool>			mIsActive;
-	std::vector<unsigned int>	componentMasks;
+	std::vector<bool>	mIsActive;
+	std::vector<size_t>	componentMasks;
 
 	std::vector<std::unique_ptr<TransformComponent>>	mTransformComponents;
 	std::vector<std::unique_ptr<CameraComponent>>		mCameraComponents;
